+++ bot/discord_api/models.py
