+++ tests/bot/discord_api/test_requests/__init__.py
@@ -1,0 +1,3 @@
+"""
+This folder has to be renamed in order to about duplications.
+"""
